Put officers in grid layout
Used bootstraps grid functionality to layout the images and text on
the page.  It takes advantage of the space available on desktops, and
scales down perfectly on mobile.  No changes have been made to the
CSS.  It's still very bare bones.  Also, added an extra officer that I
initially forgot.