Added mock instructors
The instructors are very similar to the officers structurally.  In the
original website they didn't seem to have any information about the
instructors, so I only provided names, but a title could easily be
inserted beneath just as with the officers.  Also added centering for
the names, just as I did with the officers.