Added group pictures carousel
Added the carousel to the main page.  It uses the carousel component
that is provided from BootStrap and is hard coded (but it could
probably be easily done with javascript).  There are no real
descriptions for the images right now, because I didn't want to name
them all, and I'm not sure they're even important.

Also, the last two photos, group9.jpg, and group8.jpg, don't work very
well because they are cropped in a different size.  group8.jpg
especially screws things up because it pushes whatever content is
below it down, in order to fit onto the page.  Probably best to make
sure all of these images are cropped to be the same ratio.