Added people information into main document
Took the information about the officers and instructors that was
provided and added it to the relevant section of the main page. I
modified the styling a little bit because the fonts looked very small,
but feel free to mess with the style to make it better looking.
Structurally, it should be sound the way it is.